Use uint32_t for uid/gid
The RPC protocol has 4 bytes for each of the fields, so the the IDs
must not exceed UINT32_MAX.

Unfortunately, nfs_set_uid() and nfs_set_gid() take an int parameter,
which is passed to rpc_set_uid_gid() which takes uint32_t parameters.
I verified in Wireshark that the double implicit casting results in the
expected numbers (uint32_t -> int -> uint32_t).